+++ Changelog.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Vulkan-Headers to 1.3.254 (#760)
 - Added `VK_NV_memory_decompression` device extension (#761)
 - Added `VK_GOOGLE_display_timing` device extension (#765)
+- Added `VK_ANDROID_external_memory_android_hardware_buffer` device extension (#769)
 
 ### Changed